Get the user's home directory from the passwd file
Some users will have home directories not rooted in /home; e.g. /u, or
under /var in some configurations.  Use the passwd file to lookup the
home directory.

In the case that the user doesn't have a home directory (e.g. system
user; although how they have an X session I do not know), the executable
check will be on the file `/.config/autoxkbmap`, which is unlikely to
exist.